#240 Refactor: controller 분리
