+++ src/main/java/com/codiary/backend/domain/member/dto/request/MemberRequestDTO.java
@@ -4,6 +4,10 @@ import com.codiary.backend.domain.member.entity.Member;
 import com.codiary.backend.global.apiPayload.code.status.ErrorStatus;
 import com.codiary.backend.global.apiPayload.exception.handler.MemberHandler;
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.Builder;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -12,6 +16,9 @@ import java.time.LocalDate;
 public class MemberRequestDTO {
 
     // 회원가입 요청 DTO
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+    @Builder
     public record MemberSignUpRequestDTO(
             String email,
             String password,
@@ -38,6 +45,9 @@ public class MemberRequestDTO {
     }
 
     // 로그인 요청 DTO
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+    @Builder
     public record MemberLoginRequestDTO(String email, String password) {
         public UsernamePasswordAuthenticationToken toAuthentication() {
             return new UsernamePasswordAuthenticationToken(email, password);
@@ -45,10 +55,16 @@ public class MemberRequestDTO {
     }
 
     // 프로필 이미지 요청 DTO
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+    @Builder
     public record MemberProfileImageRequestDTO(MultipartFile image) {
     }
 
     // 회원 정보 요청 DTO
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+    @Builder
     public record MemberInfoDTO(
             String birth,
             String introduction,
@@ -57,7 +73,27 @@ public class MemberRequestDTO {
             String discord) {
     }
 
+    // refresh token 요청 DTO
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+    @Builder
     public record refreshRequestDTO(
             String refreshToken) {
+    }
+
+    // 이메일 확인 요청 DTO
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+    @Builder
+    public record emailDuplicationCheckRequestDTO(
+            String email) {
+    }
+
+    // 이메일 확인 요청 DTO
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+    @Builder
+    public record nickNameDuplicationCheckRequestDTO(
+            String nickName) {
     }
 }
